feat: addin resolve async internal
