+++ tests/functional/event_handler/_pydantic/test_resolve_async_validation.py
@@ -1,0 +1,35 @@
+import asyncio
+
+from aws_lambda_powertools.event_handler.api_gateway import (
+    APIGatewayHttpResolver,
+    BaseRouter,
+)
+from tests.functional.utils import load_event
+
+API_RESTV2_EVENT = load_event("apiGatewayProxyV2Event_GET.json")
+
+
+def _setup_app(app, event):
+    BaseRouter.current_event = app._to_proxy_event(event)
+    BaseRouter.lambda_context = {}
+
+
+class TestResolveAsyncValidation:
+    def test_validation_middleware_created_and_used(self):
+        # GIVEN a resolver with validation enabled and an async handler
+        app = APIGatewayHttpResolver(enable_validation=True)
+
+        @app.get("/my/path")
+        async def get_lambda() -> dict:
+            await asyncio.sleep(0)
+            return {"message": "validated"}
+
+        # WHEN calling _resolve_async
+        _setup_app(app, API_RESTV2_EVENT)
+        result = asyncio.run(app._resolve_async())
+
+        # THEN the validation middlewares are created and the response is valid
+        response = result.build(app.current_event, app._cors)
+        assert response["statusCode"] == 200
+        assert hasattr(app, "_request_validation_middleware")
+        assert hasattr(app, "_response_validation_middleware")
